+++ tensor_test.go
@@ -16,23 +16,6 @@ func ExampleTensor() {
 	// Output:
 }
 
-func TestLogSoftmax(t *testing.T) {
-	a := assert.New(t)
-	x := torch.RandN([]int64{1, 6}, false)
-	out := x.LogSoftmax(1)
-	// TODO(yancey1989): convert torchTensor as Go slice, that we can
-	// check the value.
-	a.NotNil(out.T)
-}
-
-func TestSqueeze(t *testing.T) {
-	x := torch.RandN([]int64{2, 1, 2, 1, 2}, false)
-	y := torch.Squeeze(x)
-	assert.NotNil(t, y.T)
-	z := torch.Squeeze(x, 1)
-	assert.NotNil(t, z.T)
-}
-
 func TestTensorItem(t *testing.T) {
 	x := torch.RandN([]int64{1}, false)
 	r := x.Item()
@@ -45,28 +28,6 @@ func TestTensorDetach(t *testing.T) {
 	assert.NotNil(t, y.T)
 	initializer.Zeros(&y)
 	assert.Equal(t, float32(0.0), x.Item())
-}
-
-func TestTensorMean(t *testing.T) {
-	x := torch.RandN([]int64{2, 3}, true)
-	y := x.Mean()
-	z := y.Item()
-	assert.NotNil(t, z)
-}
-
-func TestAdd(t *testing.T) {
-	x := torch.RandN([]int64{2, 3}, false)
-	y := torch.RandN([]int64{2, 3}, false)
-	z := torch.Add(x, y, 1)
-	x.AddI(y, 1)
-	assert.True(t, torch.Equal(x, z))
-}
-
-func TestStack(t *testing.T) {
-	t1 := torch.RandN([]int64{2, 3}, false)
-	t2 := torch.RandN([]int64{2, 3}, false)
-	out := torch.Stack([]torch.Tensor{t1, t2}, 0)
-	assert.Equal(t, []int64{2, 2, 3}, out.Shape())
 }
 
 func TestFromBlob(t *testing.T) {
